rpc-core/src/main/java/me/hypocrite30/rpc/core/serialize/protostuff/ProtostuffSerializer.java
package me.hypocrite30.rpc.core.serialize.protostuff;

import io.protostuff.LinkedBuffer;
import io.protostuff.ProtostuffIOUtil;
import io.protostuff.Schema;
import io.protostuff.runtime.RuntimeSchema;
import me.hypocrite30.rpc.core.serialize.Serializer;

/**
 * Serializer by protostuff
 *
 * @Author: Hypocrite30
 * @Date: 2021/12/5 17:34
 */
public class ProtostuffSerializer implements Serializer {

    /**
     * allocate buffer in advance
     */
    private static final LinkedBuffer BUFFER = LinkedBuffer.allocate(LinkedBuffer.DEFAULT_BUFFER_SIZE);

    @Override
    public byte[] serialize(Object obj) {
        Class<?> clazz = obj.getClass();
        Schema schema = RuntimeSchema.getSchema(clazz);
        byte[] bytes;
        try {
            bytes = ProtostuffIOUtil.toByteArray(obj, schema, BUFFER);
        } finally {
            BUFFER.clear();
        }
        return bytes;
    }

    @Override
    public <T> T deserialize(byte[] bytes, Class<T> clazz) {
        Schema<T> schema = RuntimeSchema.getSchema(clazz);
        T obj = schema.newMessage();
        ProtostuffIOUtil.mergeFrom(bytes, obj, schema);
        return obj;
    }
}
